+++ src/Controller/FrontController.php
@@ -2,8 +2,10 @@
 
 namespace App\Controller;
 
+use App\Constant\UsefulLink;
 use Knp\Bundle\SnappyBundle\Snappy\Response\PdfResponse;
 use Knp\Snappy\Pdf;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
 final class FrontController extends BaseController
@@ -22,5 +24,20 @@ final class FrontController extends BaseController
             'application/pdf',
             'inline'
         );
+    }
+
+    #[Route('/liens-utiles', name: 'front_useful_links')]
+    public function usefulLinks(): Response
+    {
+        return $this->render('front/useful_links.html.twig', [
+            // To show a different approach
+            'devOpsLinks' => UsefulLink::getDevOpsLinks(),
+        ]);
+    }
+
+    #[Route('/soutien', name: 'front_support_me')]
+    public function supportMe(): Response
+    {
+        return $this->render('front/support_me.html.twig');
     }
 }