+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-11-04
+
+[Release v1.1.0](https://github.com/MathiasDaverede/job-search/releases/tag/v1.1.0)
+
+### Added
+
+- feat(ui) [#9](https://github.com/MathiasDaverede/job-search/issues/9): display version and changelog
+- feat(ui) [#10](https://github.com/MathiasDaverede/job-search/issues/10): add useful links page for programmers
+- feat(ui) [#11](https://github.com/MathiasDaverede/job-search/issues/11): add support me page
+
+### Changed
+
+- docs [#12](https://github.com/MathiasDaverede/job-search/issues/12): finalize README.md
+
 ## [1.0.0] - 2025-11-04
 
 [Release v1.0.0](https://github.com/MathiasDaverede/job-search/releases/tag/v1.0.0)